move plugins/init.tl to plugin-manager.tl
