+++ lua/legendary/plugin-manager.lua
