+++ teal/legendary/plugin-manager.tl
